Move update inventory method on Sum handler
